refactor: rename main.py to config.py and refactor its references add: test_Scrape1qxs to test the basic scraping. Also monkeypatch the src.config.db to use test db
