+++ src/celery_tasks/beat_tasks.py
@@ -1,4 +1,4 @@
-from src.main import celery_app
+from config import celery_app
 
 
 @celery_app.task
